web2.0/requirements.txt
torch==1.9.0
Pillow==8.3.1
torchvision==0.10.0
numpy==1.21.0
scikit-learn==0.24.2
pandas==1.3.0
matplotlib==3.4.2
opencv-python==4.5.3.56
